+++ woofy-backend/src/main/java/com/woofy/woofy_backend/Repositories/BusinessRepository.java
@@ -3,6 +3,6 @@ package com.woofy.woofy_backend.Repositories;
 import com.woofy.woofy_backend.Models.Entities.BusinessEntities.BusinessEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BusinessRepository extends JpaRepository<BusinessEntity, Long> {
+public interface BusinessRepository extends JpaRepository<BusinessEntity, Integer> {
 
 }
